chore: удаление игр и мёртвого кода (~1200 строк), экономика монет сохранена
По решению владельца блэкджек удаляется совсем; точка расширения для будущих
игр оставлена (комментарий в main.py у include_router + app/services/coins.py).

- Удалены: handlers/games.py (блэкджек /21, /score, /подарить — роутер и так
  не был подключён), handlers/welcome.py (не подключён), services/games.py,
  services/proactive.py (после #266 его никто не вызывал: трекер активности
  писался, но не читался).
- Экономика монет вынесена в app/services/coins.py (get_or_create_stats,
  transfer_coins) — /магазин, /доработка, /addcoins работают как раньше.
- main.py: убраны джобы send_weekly_leaderboard, check_game_timeouts,
  cleanup_blackjack_commands; помечена точка подключения будущих игр.
- config: удалён неиспользуемый ai_feature_proactive.
- Админ-справка: убраны категории «Генерация картинок» (6 команд, выпилены
  вместе с images) и «Каталог услуг» (/usluga не существует); у /addcoins
  убран несуществующий лимит «10 монет».

Модель GameState и колонки UserStat.games_played/wins оставлены — таблицы в
живой БД, /reset_stats их чистит; пригодятся будущему игровому движку.
