+++ app/utils/admin_help.py
@@ -26,7 +26,7 @@ ADMIN_CATEGORIES: dict[str, tuple[str, str]] = {
         "<i>3 страйка = автоматический мут на 24 часа</i>\n"
         "\n"
         "/addcoins &lt;кол-во&gt; — начислить монеты\n"
-        "<i>Максимум 10 монет за раз, ответьте на сообщение</i>",
+        "<i>Ответьте на сообщение пользователя</i>",
     ),
     "ai": (
         "🤖 Управление ИИ",
@@ -66,14 +66,6 @@ ADMIN_CATEGORIES: dict[str, tuple[str, str]] = {
         "/rag_sync — пересобрать всю базу знаний\n"
         "<i>Систематизирует и переиндексирует все записи</i>",
     ),
-    "svc": (
-        "🏘 Каталог услуг",
-        "🏘 <b>КАТАЛОГ УСЛУГ</b>\n"
-        "\n"
-        "/usluga — добавить услугу жителя в каталог\n"
-        "<i>Ответьте на сообщение с описанием услуги\n"
-        "Работает только в топике «Услуги»</i>",
-    ),
     "pub": (
         "📝 Публикация",
         "📝 <b>ПУБЛИКАЦИЯ</b>\n"
@@ -83,31 +75,6 @@ ADMIN_CATEGORIES: dict[str, tuple[str, str]] = {
         "\n"
         "/text — отправить текст от лица бота\n"
         "<i>Бот опубликует ваш текст как своё сообщение</i>",
-    ),
-    "img": (
-        "🖼 Генерация картинок",
-        "🖼 <b>ГЕНЕРАЦИЯ КАРТИНОК (AI)</b>\n"
-        "\n"
-        "/meme &lt;описание&gt; — мем про жизнь в ЖК\n"
-        "<i>Пример: /meme шлагбаум опять завис, соседи ждут</i>\n"
-        "\n"
-        "/poster &lt;описание&gt; — афиша мероприятия\n"
-        "<i>Пример: /poster собрание жителей в воскресенье в 19:00</i>\n"
-        "\n"
-        "/digest_image &lt;описание&gt; — иллюстрация для дайджеста\n"
-        "<i>Пример: /digest_image утренние новости ЖК</i>\n"
-        "\n"
-        "/rules_image &lt;описание&gt; — иллюстрация правила чата\n"
-        "<i>Пример: /rules_image не шумите после 23:00</i>\n"
-        "\n"
-        "/welcome_card &lt;описание&gt; — карточка приветствия новосёла\n"
-        "<i>Пример: /welcome_card добро пожаловать в наш ЖК</i>\n"
-        "\n"
-        "/warning_image &lt;описание&gt; — предупреждающее изображение\n"
-        "<i>Пример: /warning_image предупреждение о парковке</i>\n"
-        "\n"
-        "<i>Картинки генерируются моделью Google Gemini 2.5 Flash Image.\n"
-        "Доступно только администраторам. Лимит: 10 шт./день.</i>",
     ),
     "mnt": (
         "🔧 Обслуживание",
